[Chore #159082364] Write unit tests for existing functionality
