+++ requirements.txt
@@ -1,14 +1,21 @@
 astroid==2.0.1
+autopep8==1.3.5
+coverage==4.5.1
 Django==2.0.6
 django-cors-middleware==1.3.1
+django-coverage==1.2.4
 django-extensions==2.1.0
+django-nose==1.4.5
 djangorestframework==3.8.2
 djangorestframework-jwt==1.11.0
 isort==4.3.4
 lazy-object-proxy==1.3.1
 mccabe==0.6.1
+nose==1.3.7
+pycodestyle==2.4.0
 PyJWT==1.6.4
 pylint==2.0.1
 pytz==2018.5
 six==1.11.0
+typed-ast==1.1.0
 wrapt==1.10.11
